kernel calc in header file
