+++ libs/superdrops/collisionprobs/CMakeLists.txt
@@ -13,7 +13,6 @@ set(LIBRARY_OUTPUT_PATH  ${CMAKE_BINARY_DIR}/lib)
 # Add executables and create library target
 set(SOURCES
   "golovinprob.cpp"
-  "longhydroprob.cpp"
   "lowlistprob.cpp"
 )
 # must use STATIC (not(!) SHARED) lib for linking to executable if build is CUDA enabled with Kokkos
